// src/app/location/models/location.models.ts
export interface Location {
  locationId?: number
  name?: string
  description?: string
  latitude?: string
  longitude?: string
  type?: string
  isActive?: boolean
  isPopular?: true
  timeZone?: string
  knowBefore?: string
  bestTimeToVisit?: string
  currency?: string
  createdAt?: boolean
  updatedAt?: boolean
}